Fix-live-widget-#26
This makes it so the post is only added to the live comments widget if
the current view is not a preview solving #26.  Adds isPreview to
o2.options and to pageMeta.  There is a conditional check when adding an
item to the live-comments widget to prevent the post being added in a
preview.