Adds local avatar support
This was tested using https://wordpress.org/plugins/wp-user-avatar/,
which has 200,000 active installs.  This will handle local avatars that
are using get_avatar() for functionality.  Solves #14, for most cases.
There are likely edge cases where this may not work, but for the most
part, I imagine this will add support for local avatars.