Tidy up how we're using `plugins_url()`
Previously, if you installed O2 by downloading `o2-master.zip` and
installed that zip file, it would go into
`/wp-content/plugins/o2-master/o2.php` and such -- which this method of
specifying the path to assets would fail on.

By being more dynamic as to specifying what folder we're in, we'll be
more reliable however we're used.