Remove date_parse fallback
There was a fallback using date_parse and the comment/post info to try
and make a valid date by filling in bad date components with ones from the
comment/post. I was unable to devise test input that would work well with it
and it would mangle the time into some completely different time more often
than not.

This is likely because date_parse accepts a "date in format accepted by
strtotime()" (http://php.net/manual/en/function.date-parse.php), so if
strtotime is unable to parse the input accurately, date_parse will fail too.