Better time parser "Now" testing
Added a 1-second threshold for the basic "Now" test in the time parser
to account for a case where the second rolls over between the time
the test starts and the time the test ends.

Also added relative "Now" testing to handle the case where e.g. someone says
"Now" in a comment and we want a "Now" relative to when
the comment was posted