Allow up to 1-second difference in insert_comment_actions test.
There was a race-condition in the insert_comment_actions test where
a test would fail if the second changed between the start and finish
of the test.

I have modified the test to handle up to 1 full second of difference
in the checking. This should resolve that condition, and if the test
takes longer than a second something has gone horribly wrong somewhere
and we'd probably want to know that it takes over a second to insert
a comment anyways.