Remove rtlcss saveUnmodified option
That option does not skip generating a file if the file has not been modified since
last generation like I thought it did. It doesn't generate files that have the same output
as the input.

I've removed that option to be able to set up all the modules with good futureproofing,
so if the style changes in such a way that a css file would be generated the module is
already set up to enqueue the correct file.