Dropdown Actions: Change the short link action to being copied to the clipboard (#160)
* Change the short link action to being copied to the clipboard

I kept clicking on this link, which takes me to the post and the window
location is the full URL. I thought I'd see if I could change that
behaviour so the short link is copied to the clipboard. I also added the
shake animation to the icon.

* Fixing whitespace in new `onShortLinkClick` function

* Add a notification to make it clearer what the link is copied

Based on feedback that it wasn't obvious the link had been copied to the
system clipboard, I've added a notification to make it more explicit.

* Fix linting error on accessing the shortlinkCopied string

It shauld have been using dot notation, and not array access.