Use the user nicename for mentions and url slug
Although the user_nicename is created out of the user_login in most cases, there can be cases when for security reason we are setting the user_nicename to something else than the user_login. In this cas the filter widget fails to get mentions or displays a 404 when clicking on the "my posts" link.
Using the user_nicename seems the right way to build the author url.